@@ -3,26 +3,44 @@
   "version": "1.0.0",
   "description": "Project containing your PowerApps Component Framework (PCF) control.",
   "scripts": {
+    "prepare": "cd ../../.. && husky install src/components/.husky",
     "build": "pcf-scripts build",
     "clean": "pcf-scripts clean",
     "rebuild": "pcf-scripts rebuild",
-    "start": "pcf-scripts start"
+    "start": "pcf-scripts start",
+    "lint": "eslint . --debug",
+    "lint:fix": "eslint . --fix"
   },
   "dependencies": {
-    "@types/node": "^10.12.18",
-    "@types/powerapps-component-framework": "1.1.0",
-    "@types/react-truncate": "^2.3.3",
-    "antd": "^4.8.5",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "@types/node": "^16.11.6",
+    "@types/powerapps-component-framework": "1.3.3",
+    "@types/react-truncate": "^2.3.4",
+    "antd": "^4.16.13",
+    "eslint": "^7.26.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin": "^1.0.1",
+    "eslint-plugin-jsdoc": "^37.0.3",
+    "husky": "^7.0.4",
+    "lint-staged": "^11.2.6",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "react-truncate": "^2.4.0",
-    "whatwg-fetch": "^3.5.0"
+    "whatwg-fetch": "^3.6.2"
   },
   "devDependencies": {
-    "@types/react": "^16.8",
-    "@types/react-dom": "^16.8",
-    "typescript": "^3.9.9",
+    "@types/react": "^17.0",
+    "@types/react-dom": "^17.0",
+    "@typescript-eslint/eslint-plugin": "^5.3.0",
+    "@typescript-eslint/eslint-plugin-tslint": "^5.3.0",
+    "@typescript-eslint/parser": "^5.3.0",
+    "eslint-plugin-import": "^2.25.2",
+    "eslint-plugin-jsdoc": "^37.0.3",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-prefer-arrow": "^1.2.3",
+    "eslint-plugin-prettier": "^4.0.0",
     "pcf-scripts": "^1",
-    "pcf-start": "^1"
+    "pcf-start": "^1",
+    "prettier": "2.4.1",
+    "typescript": "^4.4.4"
   }
 }
