@@ -1,7 +1,7 @@
 // https://docs.microsoft.com/en-us/powerapps/developer/common-data-service/entity-metadata
 
 export enum EntityMetadataProperties {
-	EntitySetName = "EntitySetName",
-	PrimaryIdAttribute = "PrimaryIdAttribute",
-	PrimaryNameAttribute  = "PrimaryNameAttribute"
+    EntitySetName = "EntitySetName",
+    PrimaryIdAttribute = "PrimaryIdAttribute",
+    PrimaryNameAttribute = "PrimaryNameAttribute"
 }
