@@ -1,5 +1,5 @@
-import { IInputs, IOutputs } from "./generated/ManifestTypes"
-import { TagPickerBaseComponent } from "./TagPickerBaseComponent"
+import { IInputs, IOutputs } from "./generated/ManifestTypes";
+import { TagPickerBaseComponent } from "./TagPickerBaseComponent";
 
 export class TagPickerGridComponent extends TagPickerBaseComponent<IInputs, IOutputs> {
     public static readonly BodyContainerDataId = "data-set-body-container";
@@ -18,19 +18,25 @@ export class TagPickerGridComponent extends TagPickerBaseComponent<IInputs, IOut
     /**
      * Used to initialize the control instance. Controls can kick off remote server calls and other initialization actions here.
      * Data-set values are not initialized here, use updateView.
+     *
      * @param context The entire property bag available to control via Context Object; It contains values as set up by the customizer mapped to property names defined in the manifest, as well as utility functions.
      * @param notifyOutputChanged A callback method to alert the framework that the control has new outputs ready to be retrieved asynchronously.
      * @param state A piece of data that persists in one session for a single user. Can be set at any point in a controls life cycle by calling 'setControlState' in the Mode interface.
      * @param container If a control is marked control-type='standard', it will receive an empty div element within which it can render its content.
      */
-    public init(context: ComponentFramework.Context<IInputs>, notifyOutputChanged: () => void, state: ComponentFramework.Dictionary, container:HTMLDivElement) {
+    public async init(
+        context: ComponentFramework.Context<IInputs>,
+        notifyOutputChanged: () => void,
+        state: ComponentFramework.Dictionary,
+        container: HTMLDivElement
+    ): Promise<void> {
         this.relatedEntity = context.parameters.relatedEntity.raw || "";
         this.relationshipEntity = context.parameters.relationshipEntity.raw || "";
         this.relationshipName = context.parameters.relationshipName.raw || "";
         this.labelText = context.parameters.labelText.raw || "";
         this.viewId = context.parameters.viewId.raw || "";
 
-        super.init(context, notifyOutputChanged, state, container);
+        await super.init(context, notifyOutputChanged, state, container);
     }
 
     public updateView(context: ComponentFramework.Context<IInputs>): void {
@@ -42,12 +48,14 @@ export class TagPickerGridComponent extends TagPickerBaseComponent<IInputs, IOut
     public destroy(): void {
         super.destroy();
 
-        if (this.observer !== undefined)
+        if (this.observer !== undefined) {
             this.observer.disconnect();
+        }
     }
 
     /**
      * It is called by the framework prior to a control receiving new data.
+     *
      * @returns an object based on nomenclature defined in manifest, expecting object[s] for property marked as “bound” or “output”
      */
     public getOutputs(): IOutputs {
@@ -61,8 +69,9 @@ export class TagPickerGridComponent extends TagPickerBaseComponent<IInputs, IOut
         this.container.classList.add(TagPickerGridComponent.ContainerClass);
 
         // exit if the observer is already configured
-        if (this.observer !== undefined)
+        if (this.observer !== undefined) {
             return;
+        }
 
         const bodyContainer = this.getBodyContainer();
 
@@ -79,8 +88,9 @@ export class TagPickerGridComponent extends TagPickerBaseComponent<IInputs, IOut
             this.observer = new MutationObserver((mutations, observer) => {
                 observer.disconnect();
 
-                if (!bodyContainer.classList.contains(TagPickerGridComponent.BodyContainerClass))
+                if (!bodyContainer.classList.contains(TagPickerGridComponent.BodyContainerClass)) {
                     bodyContainer.classList.add(TagPickerGridComponent.BodyContainerClass);
+                }
 
                 observer.observe(bodyContainer, options);
             });
@@ -104,8 +114,9 @@ export class TagPickerGridComponent extends TagPickerBaseComponent<IInputs, IOut
             const dataId = parent.getAttribute("data-id");
             parent = parent.parentElement;
 
-            if (dataId === TagPickerGridComponent.BodyContainerDataId)
+            if (dataId === TagPickerGridComponent.BodyContainerDataId) {
                 break;
+            }
         }
 
         return parent;
