@@ -17,16 +17,16 @@ import {
     Selection,
     IDetailsListCheckboxProps,
     IDetailsRowProps
-} from 'office-ui-fabric-react';
-import { initializeIcons } from '@uifabric/icons';
-import { IconButton } from '@fluentui/react/lib/Button';
-import { SharedColors } from '@fluentui/theme';
+} from "office-ui-fabric-react";
+import { initializeIcons } from "@uifabric/icons";
+import { IconButton } from "@fluentui/react/lib/Button";
+import { SharedColors } from "@fluentui/theme";
 import { IInputs } from "./generated/ManifestTypes";
 import { ITaskItem, ITaskManagerProps, ITaskManagerState } from "./interfaces";
 import { TaskStatus } from "./enums";
 initializeIcons();
 
-export class TaskManager extends React.Component<ITaskManagerProps, ITaskManagerState>{
+export class TaskManager extends React.Component<ITaskManagerProps, ITaskManagerState> {
     private _root = React.createRef<IDetailsList>();
     private _columns: IColumn[];
     private _context: ComponentFramework.Context<IInputs>;
@@ -40,45 +40,35 @@ export class TaskManager extends React.Component<ITaskManagerProps, ITaskManager
 
         this._context = props.context;
         this._selection = new Selection({
-            onSelectionChanged: this.handleOnSelectionChanged
+            onSelectionChanged: () => this.handleOnSelectionChanged()
         });
 
         this.state = {
             tasks: props.tasks,
             selectedItems: this._selection.getSelection(),
-            showInactive: false,
+            showInactive: false
         };
 
         // Define columns
-        this._columns = [{
-            key: 'name',
-            name: this._context.resources.getString("label_name"),
-            fieldName: 'name',
-            minWidth: 100,
-            isResizable: true
-        }];
+        this._columns = [
+            {
+                key: "name",
+                name: this._context.resources.getString("label_name"),
+                fieldName: "name",
+                minWidth: 100,
+                isResizable: true
+            }
+        ];
     }
 
-    public componentDidUpdate(prevProps: ITaskManagerProps) {
+    public componentDidUpdate(prevProps: ITaskManagerProps): void {
         // If props are different it means changes are not tracked in the state
         if (this.props.tasks !== prevProps.tasks) {
-            this.setState((prevState, newProps) => {
-                return { tasks: newProps.tasks, selectedItems: [] };
-            });
+            this.setState((prevState, newProps) => ({ tasks: newProps.tasks, selectedItems: [] }));
         }
     }
 
-
-    private bindThis() {
-        this.handleRenderColumn = this.handleRenderColumn.bind(this);
-        this.handleShowAllClick = this.handleShowAllClick.bind(this);
-        this.handleAddOnClick = this.handleAddOnClick.bind(this);
-        this.handleDeleteTask = this.handleDeleteTask.bind(this);
-        this.handleRenderRow = this.handleRenderRow.bind(this);
-        this.handleOnSelectionChanged = this.handleOnSelectionChanged.bind(this);
-    }
-
-    public render() {
+    public render(): JSX.Element {
         const buttonStyles: Partial<IButtonStyles> = {
             icon: {
                 color: SharedColors.gray40
@@ -90,8 +80,20 @@ export class TaskManager extends React.Component<ITaskManagerProps, ITaskManager
                 <div style={{ display: "flex", padding: "10px 10px 0 10px" }}>
                     <h3>{this.props.panelTitle}</h3>
                     <div style={{ marginLeft: "auto" }}>
-                        <IconButton iconProps={{ iconName: 'Add' }} title={this._context.resources.getString("label_addtask")} ariaLabel={this._context.resources.getString("label_addtask")} styles={buttonStyles} onClick={this.handleAddOnClick} />
-                        <IconButton iconProps={{ iconName: 'AllApps' }} title={this._context.resources.getString("label_toggletaskvisibility")} ariaLabel={this._context.resources.getString("label_toggletaskvisibility")} styles={buttonStyles} onClick={this.handleShowAllClick} />
+                        <IconButton
+                            iconProps={{ iconName: "Add" }}
+                            title={this._context.resources.getString("label_addtask")}
+                            ariaLabel={this._context.resources.getString("label_addtask")}
+                            styles={buttonStyles}
+                            onClick={() => this.handleAddOnClick()}
+                        />
+                        <IconButton
+                            iconProps={{ iconName: "AllApps" }}
+                            title={this._context.resources.getString("label_toggletaskvisibility")}
+                            ariaLabel={this._context.resources.getString("label_toggletaskvisibility")}
+                            styles={buttonStyles}
+                            onClick={() => this.handleShowAllClick()}
+                        />
                     </div>
                 </div>
                 <DetailsList
@@ -102,12 +104,12 @@ export class TaskManager extends React.Component<ITaskManagerProps, ITaskManager
                     columns={this._columns}
                     ariaLabelForSelectionColumn={this._context.resources.getString("label_toggletaskcompleted")}
                     checkButtonAriaLabel={this._context.resources.getString("label_toggletaskcompleted")}
-                    onRenderDetailsHeader={this.handleRenderDetailsHeader}
+                    onRenderDetailsHeader={(props, defaultrender) => this.handleRenderDetailsHeader(props, defaultrender)}
                     isHeaderVisible={false}
                     checkboxVisibility={CheckboxVisibility.always}
-                    onRenderItemColumn={this.handleRenderColumn}
-                    onRenderRow={this.handleRenderRow}
-                    onRenderCheckbox={this.handleRenderCheckbox}
+                    onRenderItemColumn={(item: ITaskItem, index, column) => this.handleRenderColumn(item, index, column)}
+                    onRenderRow={props => this.handleRenderRow(props)}
+                    onRenderCheckbox={props => this.handleRenderCheckbox(props)}
                     selection={this._selection}
                     selectionPreservedOnEmptyClick={true}
                     selectionMode={SelectionMode.multiple}
@@ -116,30 +118,39 @@ export class TaskManager extends React.Component<ITaskManagerProps, ITaskManager
         );
     }
 
-    private handleOnSelectionChanged() {
+    private bindThis() {
+        this.handleRenderColumn = this.handleRenderColumn.bind(this);
+        this.handleShowAllClick = this.handleShowAllClick.bind(this);
+        this.handleAddOnClick = this.handleAddOnClick.bind(this);
+        this.handleDeleteTask = this.handleDeleteTask.bind(this);
+        this.handleRenderRow = this.handleRenderRow.bind(this);
+        this.handleOnSelectionChanged = this.handleOnSelectionChanged.bind(this);
+    }
+
+    private handleOnSelectionChanged(): void {
         try {
             // Isolate checked and unchecked item delta between state and current UI
             const currentSelection = this._selection.getSelection();
             const checked = currentSelection.filter(t => !this.state.selectedItems.find(s => s.key === t.key));
             const unchecked = this.state.selectedItems.filter(t => !currentSelection.find(s => s.key === t.key));
 
-            this.setState((prevState, props) => {
-                return { selectedItems: this._selection.getSelection() }
-            });
+            this.setState((prevState, props) => ({ selectedItems: this._selection.getSelection() }));
 
             if (checked.length > 0) {
-                this._context.webAPI.updateRecord("task", checked[0].key as string, { statecode: 1, statuscode: TaskStatus.Completed })
+                this._context.webAPI
+                    .updateRecord("task", checked[0].key as string, { statecode: 1, statuscode: TaskStatus.Completed })
+                    .catch(ex => console.error(ex));
+            } else if (unchecked.length > 0) {
+                this._context.webAPI
+                    .updateRecord("task", unchecked[0].key as string, { statecode: 0, statuscode: TaskStatus.InProgress })
                     .catch(ex => console.error(ex));
             }
-            else if (unchecked.length > 0) {
-                this._context.webAPI.updateRecord("task", unchecked[0].key as string, { statecode: 0, statuscode: TaskStatus.InProgress })
-                    .catch(ex => console.error(ex));
-            }
-        } catch (e) { console.error(e); }
+        } catch (e) {
+            console.error(e);
+        }
     }
 
-    private handleAddOnClick() {
-
+    private handleAddOnClick(): void {
         // Get page context to pass in to quick create
         // NOTE: There appear to be a mismatch between EntityReference referred in context object and what's defined in XrmDefinitelyTyped. Interestingly only the latter works.
         const pageContext = (this.props.context as any).mode.contextInfo;
@@ -149,11 +160,15 @@ export class TaskManager extends React.Component<ITaskManagerProps, ITaskManager
             entityType: pageContext.entityTypeName
         };
 
-        this.props.context?.navigation.openForm({
-            entityName: "task",
-            useQuickCreateForm: true,
-            createFromEntity: entityRef
-        }, undefined)
+        this.props.context?.navigation
+            .openForm(
+                {
+                    entityName: "task",
+                    useQuickCreateForm: true,
+                    createFromEntity: entityRef
+                },
+                undefined
+            )
             .then(
                 taskRef => {
                     // Component should magically refresh
@@ -164,22 +179,33 @@ export class TaskManager extends React.Component<ITaskManagerProps, ITaskManager
             );
     }
 
-    private handleShowAllClick() {
-        this.setState((prevState, props) => {
-            return { showInactive: !prevState.showInactive }
-        });
+    private handleShowAllClick(): void {
+        this.setState((prevState, props) => ({ showInactive: !prevState.showInactive }));
     }
 
     private handleRenderRow(props?: IDetailsRowProps) {
         // If rendering an inactive task do not show checkbox
-        if(props && !(props.item as ITaskItem).isActive){
+        if (props && !(props.item as ITaskItem).isActive) {
             props.checkboxVisibility = CheckboxVisibility.hidden;
-            return <DetailsRow data-selection-disabled={true} {...props} styles={{ root: { alignItems: "center" } }} rowFieldsAs={this.handleRenderRowFields} />
+            return (
+                <DetailsRow
+                    data-selection-disabled={true}
+                    {...props}
+                    styles={{ root: { alignItems: "center" } }}
+                    rowFieldsAs={(fieldProps: IDetailsRowFieldsProps) => this.handleRenderRowFields(fieldProps)}
+                />
+            );
         }
-        return props ? <DetailsRow {...props} styles={{ root: { alignItems: "center" } }} rowFieldsAs={this.handleRenderRowFields} /> : null;
-    };
+        return props ? (
+            <DetailsRow
+                {...props}
+                styles={{ root: { alignItems: "center" } }}
+                rowFieldsAs={(fieldProps: IDetailsRowFieldsProps) => this.handleRenderRowFields(fieldProps)}
+            />
+        ) : null;
+    }
 
-    private handleRenderRowFields(props: IDetailsRowFieldsProps) {
+    private handleRenderRowFields(props: IDetailsRowFieldsProps): JSX.Element {
         return (
             // BUG: Not perfect here, there seems to be a single pixel that allows selection
             <span data-selection-disabled={true}>
@@ -188,45 +214,74 @@ export class TaskManager extends React.Component<ITaskManagerProps, ITaskManager
         );
     }
 
-    private handleRenderCheckbox(props?: IDetailsListCheckboxProps) {
+    private handleRenderCheckbox(props?: IDetailsListCheckboxProps): JSX.Element {
         return (
-            <div style={{ pointerEvents: 'none' }}>
+            <div style={{ pointerEvents: "none" }}>
                 <Checkbox checked={props?.checked} />
             </div>
         );
     }
 
-    private handleRenderDetailsHeader(props?: IDetailsHeaderProps, _defaultRender?: IRenderFunction<IDetailsHeaderProps>) {
-        return <DetailsHeader {...(props ?? { columns: undefined, selection: undefined, selectMode: undefined, layoutMode: DetailsListLayoutMode.fixedColumns })} ariaLabelForToggleAllGroupsButton={'Expand collapse groups'} />;
-    }
-
-    private handleRenderColumn(item?: ITaskItem, index?: number, column?: IColumn) {
-        return <div className="task-wrapper">
-            <div className="task-content">
-                <span className="task-title">{item?.subject}</span>
-                {this.props.badgeConfig?.map((badgeConfigItem) => {
-                    // Cast as any to access property value from variable name which represent the attribute name
-                    const optionKey = (item as any)[badgeConfigItem.name];
-                    const optionMetadata = badgeConfigItem.values?.find(v => v.key == optionKey);
-
-                    // This transforms all badge configurations into <span> elements if the current value matches something in the configuration.
-                    // If the value was not mapped in the configuration, don't do anything with it.
-                    if (optionMetadata) {
-                        return <span className="badge" style={{ backgroundColor: optionMetadata.color ?? "gray" }} key={item?.key + "badge" + optionMetadata.label}>{optionMetadata.label}</span>
-                    }
+    private handleRenderDetailsHeader(props?: IDetailsHeaderProps, _defaultRender?: IRenderFunction<IDetailsHeaderProps>): JSX.Element {
+        return (
+            <DetailsHeader
+                {...(props ?? {
+                    columns: undefined,
+                    selection: undefined,
+                    selectMode: undefined,
+                    layoutMode: DetailsListLayoutMode.fixedColumns
                 })}
-                <div className="task-description">{item?.description}</div>
-            </div>
-            <div className="task-action">
-                {item?.isActive && <IconButton iconProps={{ iconName: 'Delete' }} title={this._context.resources.getString("label_canceltask")} ariaLabel={this._context.resources.getString("label_canceltask")} styles={{ icon: { color: SharedColors.red10 } }} onClick={() => this.handleDeleteTask(item?.key)} />}
-            </div>
-        </div>;
-        //RemoveFromTrash
+                ariaLabelForToggleAllGroupsButton={"Expand collapse groups"}
+            />
+        );
     }
 
-    private handleDeleteTask(taskid?: string) {
+    private handleRenderColumn(item?: ITaskItem, index?: number, column?: IColumn): JSX.Element {
+        return (
+            <div className="task-wrapper">
+                <div className="task-content">
+                    <span className="task-title">{item?.subject}</span>
+                    {this.props.badgeConfig?.map(badgeConfigItem => {
+                        // Cast as any to access property value from variable name which represent the attribute name
+                        const optionKey = (item as any)[badgeConfigItem.name];
+                        const optionMetadata = badgeConfigItem.values?.find(v => v.key === optionKey);
+
+                        // This transforms all badge configurations into <span> elements if the current value matches something in the configuration.
+                        // If the value was not mapped in the configuration, don't do anything with it.
+                        if (optionMetadata) {
+                            return (
+                                <span
+                                    className="badge"
+                                    style={{ backgroundColor: optionMetadata.color ?? "gray" }}
+                                    key={(item?.key as string) + "badge" + optionMetadata.label}
+                                >
+                                    {optionMetadata.label}
+                                </span>
+                            );
+                        }
+                    })}
+                    <div className="task-description">{item?.description}</div>
+                </div>
+                <div className="task-action">
+                    {item?.isActive && (
+                        <IconButton
+                            iconProps={{ iconName: "Delete" }}
+                            title={this._context.resources.getString("label_canceltask")}
+                            ariaLabel={this._context.resources.getString("label_canceltask")}
+                            styles={{ icon: { color: SharedColors.red10 } }}
+                            onClick={() => this.handleDeleteTask(item?.key)}
+                        />
+                    )}
+                </div>
+            </div>
+        );
+        // RemoveFromTrash
+    }
+
+    private handleDeleteTask(taskid?: string): void {
         if (taskid) {
-            this._context.webAPI.updateRecord("task", taskid, { statecode: 2, statuscode: TaskStatus.Canceled })
+            this._context.webAPI
+                .updateRecord("task", taskid, { statecode: 2, statuscode: TaskStatus.Canceled })
                 .then(r => {
                     this.setState((prevState, props) => {
                         // Clone array, update deleted item to inactive then update state which will re-render.
@@ -236,7 +291,7 @@ export class TaskManager extends React.Component<ITaskManagerProps, ITaskManager
                             task.isActive = false;
                             task.statuscode = TaskStatus.Canceled;
                         }
-                        return { tasks: clonedItems }
+                        return { tasks: clonedItems };
                     });
                 })
                 .catch(ex => console.error(ex));
